ajout de saync trait et doonc enregistrement de vols qui permet donc de simplment ajouter la reuqte ogn dans vec::du
