+++ Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = "0.1.72"
 chrono = "0.4.24"
 env_logger = "0.10.0"
 hyper = { version = "0.14", features = ["full"] }
